requirements.txt
jupyter==1.0.0,
numpy==1.18.5,
pandas~=1.1.3,
opencv-python~=4.1.2,
omegaconf==1.4.1,
hydra-core==0.11.3,
pytorch-lightning==0.8.1,
albumentations==0.4.5,
Pillow==7.0.0,
pycocotools==2.0,
matplotlib==3.2.2,
torchvision==0.7.0,
tqdm==4.45.0,
torch==1.6.0,
torchtext==0.3.1,
torchsummary==1.5.1